STM32F407 LED BLİNK
When I add and enter the ascii codes of my name and surname, we get the mod4 of the numbers that come separately.
We choose the leds according to the number in mod4. Afterwards, the selected leds light up first as the name and then as the surname. The number of leds is as much as the letter number of my name and surname.